backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.16.5",
    "asyncpg>=0.30.0",
    "fastapi>=0.116.1",
    "google-genai>=1.33.0",
    "greenlet>=3.2.4",
    "httpx>=0.28.1",
    "psycopg>=3.2.9",
    "pydantic[email]>=2.11.7",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.1.0",
    "pytest-postgresql>=7.0.2",
    "ruff>=0.12.12",
    "sqlalchemy>=2.0.43",
    "uvicorn[standard]>=0.35.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
pythonpath = "."

[tool.ruff]
# Set the target Python version for your project.
target-version = "py313"

# Exclude common directories and files from linting.
exclude = [
    ".venv",
    "__pycache__",
    "build",
    "dist"
]

line-length = 200

[tool.ruff.lint]
# Enable a basic but comprehensive set of rules.
# F: Pyflakes, E: pycodestyle, I: isort.
select = ["F", "E", "I"]

# Set fixable rules to ALL, so you can run `ruff --fix`.
fixable = ["ALL"]
